Add files via upload
Added a description box.
Started setting the structure for classes and races. They will be implementations of an IClass interface. A CharData container will have one of these as a member.
Began linking the selected class in the class selection box to a description, shown in the top right of the New Character window.