Added race images
These are pulled from the database instead of directly from a file. This makes it easy for end users to changes, and easier for us to manage. This was also a horrible pain the neck that I spent days banging my head against my keyboard trying to make happen. This is also part of the goal we are working towards, which is the program pulling everything from the database. This would allow custom classes, races, and more. Stay tuned!